Edit home page
Added a new image and location div to the homepage. Opening hours, location and images are in the info div. The layout of the info div is done with a grid.

The homepage code was organized in a cleaner way, where each section is easy to find and edit.
In addition, the function that is now exported from the module does not contain the entire module code, but only adds the necessary sections to the content div.